GS-13502 Identified another participant with the same name for Space.
Ephemeral node of Space instance hasn't yet expired,
before a new instantiation has began, leading to subsequent failures
while trying to instantiate.

When the first instantiation fails, the ZookeeperAttributeStore is not
closed properly, adding more and more ephemeral node registrations to
the ZooKeeper.

Solution:
1. Add proper termination for Curator client for
ZookeeperAttributeStore and ZooKeeperBasedGSMLeaderSelectorHandler

2. When a matching participant is identified, retry after session timeout
to make sure it is not a stale participant.